backend/rag_engine/requirements.txt
fastapi
uvicorn
newspaper3k
chromadb
sentence-transformers
pydantic
openai
requests
python-dotenv
beautifulsoup4
lxml[html_clean]
